update readme, move client to its own file
