+++ README.md
@@ -431,7 +431,6 @@ In Gateway mode the following options are not allowed (the plugin will throw an 
 * `schema`
 * `resolvers`
 * `loaders`
-* `subscription`
 
 Also, using the following decorator methods will throw:
 * `app.graphql.defineResolvers`
@@ -502,6 +501,7 @@ __fastify-gql__ supports the following options:
   * `gateway.services`: Service[] An array of GraphQL services that are part of the gateway
     * `service.name`: A unique name for the service. Required.
     * `service.url`: The url of the service endpoint. Required
+    * `service.wsUrl`: The url of the websocket endpoint
     * `service.rewriteHeaders`: `Function` A function that gets the original headers as a parameter and returns an object containing values that should be added to the headers
 
 #### queryDepth example
