+++ lib/gateway/request.js
@@ -27,7 +27,6 @@ function buildRequest (opts) {
   }
 
   function close () {
-    console.log(opts.url, 'close')
     agents['http:'].destroy()
     agents['https:'].destroy()
   }
