+++ lib/gateway/service-map.js
@@ -1,7 +1,6 @@
 'use strict'
 const pmap = require('p-map')
 const URL = require('url').URL
-const WebSocket = require('ws')
 const {
   isTypeExtensionNode,
   isTypeDefinitionNode,
@@ -9,6 +8,7 @@ const {
 } = require('graphql')
 
 const { buildRequest, sendRequest } = require('./request')
+const createClient = require('./socket-connection')
 
 function hasExternalDirective (field) {
   return field.directives.some(directive => directive.name.value === 'external')
@@ -86,39 +86,20 @@ async function buildServiceMap (services, subscriber) {
     }
 
     if (service.wsUrl) {
-      const ws = new WebSocket(service.wsUrl, 'graphql-ws')
-      const client = WebSocket.createWebSocketStream(ws, { encoding: 'utf8', objectMode: true })
-      client.setEncoding('utf8')
-      client.write(JSON.stringify({
-        type: 'connection_init'
-      }))
-      client.on('data', async (chunk) => {
-        const data = JSON.parse(chunk)
-        if (data.type === 'data') {
-          await subscriber.publish({
-            topic: `${service.name}_${data.id}`,
-            payload: data.payload.data
-          })
-        }
-      })
-      client.on('error', (e) => {
-        // TODO add proper error handling
-        console.log('PANIC some error', e)
-        throw e
-      })
-      ws.on('close', () => {
-        console.log('ws close')
-        client.end()
-      })
+      const client = createClient(service.wsUrl, subscriber, service.name)
 
-      serviceConfig.subscriptionQueryMap = new Set()
-      serviceConfig.createSubscription = (query, args = {}) => {
-        // TODO find a better way to generate unique id base on query and args
-        const id = (query + '_' + JSON.stringify(args)).replace(/\s/gi, '') // cannot contain spaces because then it fails don't know why
+      const subscriptionQueryMap = {}
+      let idCount = 0
 
-        if (serviceConfig.subscriptionQueryMap.has(id)) {
+      serviceConfig.createSubscription = (query, args) => {
+        const subscriptionString = JSON.stringify({ query, args })
+        let id = subscriptionQueryMap[subscriptionString]
+
+        if (id) {
           return id
         }
+
+        id = ++idCount
 
         client.write(JSON.stringify({
           id,
@@ -129,7 +110,7 @@ async function buildServiceMap (services, subscriber) {
           }
         }))
 
-        serviceConfig.subscriptionQueryMap.add(id)
+        subscriptionQueryMap[subscriptionString] = id
 
         return id
       }
